+++ src/app/admin/components/user/edit.controller.js
@@ -107,12 +107,10 @@
                 vm.roles.push('ROLE_ONC_STAFF');
             }
             if (!vm.atlId) {
-                vm.roles.push('ROLE_ACB_ADMIN');
-                vm.roles.push('ROLE_ACB_STAFF');
+                vm.roles.push('ROLE_ACB');
             }
             if (!vm.acbId) {
-                vm.roles.push('ROLE_ATL_ADMIN');
-                vm.roles.push('ROLE_ATL_STAFF');
+                vm.roles.push('ROLE_ATL');
             }
         }